requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
greenlet==3.0.1
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
playwright==1.40.0
pyee==11.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
undetected-playwright-patch==1.40.0.post1700587210000
urllib3==2.2.3
